requirements.txt
azure-functions==1.17.0
install==1.3.5
load-dotenv==0.1.0
numpy==1.26.2
pandas==2.1.4
polars==0.19.19
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3